+++ central/src/auth/mod.rs
@@ -1,4 +1,3 @@
 mod keycloak;
 pub(crate) mod config;
 mod authentik;
-mod authentik_test;
